+++ src/components/section/SectionComponent.tsx
@@ -1,10 +1,4 @@
-import {
-  HTMLAttributes,
-  FC,
-  PropsWithChildren,
-  forwardRef,
-  useRef,
-} from "react";
+import { HTMLAttributes, FC, PropsWithChildren, forwardRef } from "react";
 import classnames from "classnames";
 
 import {
@@ -20,23 +14,24 @@ export const Section = forwardRef<
   HTMLDivElement,
   SectionProps & Omit<HTMLAttributes<HTMLElement>, "style" | "content">
 >(
-  ({
-    headline,
-    content,
-    headerSpacing,
-    width = "default",
-    style = "default",
-    spotlight = false,
-    backgroundColor = "default",
-    spaceBefore = "default",
-    spaceAfter = "default",
-    className,
-    inverted,
-    buttons = [],
-    ...props
-  }) => {
-    const parentRef = useRef();
-
+  (
+    {
+      headline,
+      content,
+      headerSpacing,
+      width = "default",
+      style = "default",
+      spotlight = false,
+      backgroundColor = "default",
+      spaceBefore = "default",
+      spaceAfter = "default",
+      className,
+      inverted,
+      buttons = [],
+      ...props
+    },
+    ref
+  ) => {
     const { large: headlineLarge = false, ...headlineRest } = {
       align: "left",
       ...headline,
@@ -85,7 +80,7 @@ export const Section = forwardRef<
         spaceBefore={spaceBefore}
         spaceAfter={spaceAfter}
         inverted={inverted}
-        ref={parentRef}
+        ref={ref}
       >
         {props.children}
       </SectionContextDefault>